+++ packages/chrome-extension/src/app/components/pr/PrEditorsApp.tsx
@@ -15,34 +15,34 @@
  */
 
 import * as React from "react";
-import { useContext, useRef, useState } from "react";
-import * as ReactDOM from "react-dom";
-import { IsolatedEditor } from "../common/IsolatedEditor";
-import { PrToolbar } from "./PrToolbar";
-import { IsolatedEditorContext } from "../common/IsolatedEditorContext";
-import { getOriginalFilePath, getUnprocessedFilePath, GitHubDomElementsPr } from "./GitHubDomElementsPr";
+import { useContext, useState } from "react";
+import { getOriginalFilePath, getUnprocessedFilePath } from "./GitHubDomElementsPr";
 import { GlobalContext } from "../common/GlobalContext";
-import { FileStatusOnPr } from "./FileStatusOnPr";
-import {
-  useEffectAfterFirstRender,
-  useEffectWithDependencies,
-  useInitialAsyncCallEffect,
-  useIsolatedEditorTogglingEffect
-} from "../common/customEffects";
+import { useEffectWithDependencies } from "../common/customEffects";
 import { Router } from "@kogito-tooling/core-api";
 import * as dependencies__ from "../../dependencies";
 import { ResolvedDomDependency } from "../../dependencies";
-import { Feature } from "../common/Feature";
-import { IsolatedEditorRef } from "../common/IsolatedEditorRef";
+import { IsolatedPrEditor } from "./IsolatedPrEditor";
 
 export function PrEditorsApp() {
-  const prFileElements = () => dependencies__.all.supportedPrFileContainers().map(e => ({ name: "", element: e }));
+  const prFileElements = () => dependencies__.array.supportedPrFileContainers().map(e => ({ name: "", element: e }));
 
   const globalContext = useContext(GlobalContext);
   const [containers, setContainers] = useState(supportedPrFileElements(prFileElements, globalContext.router));
 
   useMutationObserverEffect(
-    newPrFileContainersMutationObserver(prFileElements, containers, setContainers, globalContext.router),
+    new MutationObserver(mutations => {
+      const addedNodes = mutations.reduce((l, r) => [...l, ...Array.from(r.addedNodes)], []);
+
+      if (addedNodes.length <= 0) {
+        return;
+      }
+
+      const newContainers = supportedPrFileElements(prFileElements, globalContext.router);
+      if (newContainers.length !== containers.length) {
+        setContainers(newContainers);
+      }
+    }),
     {
       childList: true,
       subtree: true
@@ -52,138 +52,10 @@ export function PrEditorsApp() {
   return (
     <>
       {containers.map(e => (
-        <IsolatedPrEditor key={getUnprocessedFilePath(e)} container={e} />
+        <IsolatedPrEditor key={getUnprocessedFilePath(e)} container={e} fileExtension={getFileExtension(e)} />
       ))}
     </>
   );
-}
-
-function IsolatedPrEditor(props: { container: ResolvedDomDependency }) {
-  const globalContext = useContext(GlobalContext);
-  const githubDomElements = new GitHubDomElementsPr(props.container, globalContext.router);
-
-  const [showOriginal, setShowOriginal] = useState(false);
-  const [isTextMode, setTextMode] = useState(true);
-  const [fileStatusOnPr, setFileStatusOnPr] = useState(FileStatusOnPr.UNKNOWN);
-
-  useIsolatedEditorTogglingEffect(isTextMode, c => githubDomElements.iframeContainer(c), props.container);
-
-  useInitialAsyncCallEffect(() => discoverFileStatusOnPr(githubDomElements), setFileStatusOnPr);
-
-  useEffectAfterFirstRender(
-    () => {
-      getFileContents().then(c => {
-        if (ref.current) {
-          ref.current.setContent(c || "");
-        }
-      });
-    },
-    [showOriginal]
-  );
-
-  const getFileContents =
-    showOriginal || fileStatusOnPr === FileStatusOnPr.DELETED
-      ? () => githubDomElements.getOriginalFileContents()
-      : () => githubDomElements.getFileContents();
-
-  const shouldAddLinkToOriginalFile =
-    fileStatusOnPr === FileStatusOnPr.CHANGED || fileStatusOnPr === FileStatusOnPr.DELETED;
-
-  const ref = useRef<IsolatedEditorRef>(null);
-
-  return (
-    <IsolatedEditorContext.Provider value={{ textMode: isTextMode, fullscreen: false }}>
-      {shouldAddLinkToOriginalFile && (
-        <Feature
-          name={"Link to original file"}
-          dependencies={deps => ({ container: () => deps.all.viewOriginalFileLinkContainer(props.container) })}
-          component={deps =>
-            ReactDOM.createPortal(
-              <a className={"pl-5 dropdown-item btn-link"} href={githubDomElements.viewOriginalFileHref()}>
-                View original file
-              </a>,
-              githubDomElements.viewOriginalFileLinkContainer(deps.container)
-            )
-          }
-        />
-      )}
-
-      <Feature
-        name={"Toolbar"}
-        dependencies={deps => ({ container: () => deps.common.toolbarContainerTarget(props.container) })}
-        component={deps =>
-          ReactDOM.createPortal(
-            <PrToolbar
-              fileStatusOnPr={fileStatusOnPr}
-              textMode={isTextMode}
-              originalDiagram={showOriginal}
-              toggleOriginal={() => setShowOriginal(prev => !prev)}
-              closeDiagram={() => setTextMode(true)}
-              onSeeAsDiagram={() => setTextMode(false)}
-            />,
-            githubDomElements.toolbarContainer(deps.container)
-          )
-        }
-      />
-
-      {ReactDOM.createPortal(
-        <IsolatedEditor
-          ref={ref}
-          textMode={isTextMode}
-          getFileContents={getFileContents}
-          openFileExtension={getFileExtension(props.container)}
-          readonly={true}
-          keepRenderedEditorInTextMode={false}
-        />,
-        githubDomElements.iframeContainer(props.container)
-      )}
-    </IsolatedEditorContext.Provider>
-  );
-}
-
-async function discoverFileStatusOnPr(githubDomElements: GitHubDomElementsPr) {
-  const hasOriginal = await githubDomElements.getOriginalFileContents();
-  const hasModified = await githubDomElements.getFileContents();
-
-  if (hasOriginal && hasModified) {
-    return FileStatusOnPr.CHANGED;
-  }
-
-  if (hasOriginal) {
-    return FileStatusOnPr.DELETED;
-  }
-
-  if (hasModified) {
-    return FileStatusOnPr.ADDED;
-  }
-
-  throw new Error("Impossible status for file on PR");
-}
-
-function getFileExtension(container: ResolvedDomDependency) {
-  return getOriginalFilePath(container)
-    .split(".")
-    .pop()!;
-}
-
-function newPrFileContainersMutationObserver(
-  prFileElements: () => ResolvedDomDependency[],
-  containers: ResolvedDomDependency[],
-  setContainers: (e: ResolvedDomDependency[]) => void,
-  router: Router
-) {
-  return new MutationObserver(mutations => {
-    const addedNodes = mutations.reduce((l, r) => [...l, ...Array.from(r.addedNodes)], []);
-
-    if (addedNodes.length <= 0) {
-      return;
-    }
-
-    const newContainers = supportedPrFileElements(prFileElements, router);
-    if (newContainers.length !== containers.length) {
-      setContainers(newContainers);
-    }
-  });
 }
 
 function supportedPrFileElements(prFileElements: () => ResolvedDomDependency[], router: Router) {
@@ -192,7 +64,7 @@ function supportedPrFileElements(prFileElements: () => ResolvedDomDependency[], 
 
 function useMutationObserverEffect(observer: MutationObserver, options: MutationObserverInit) {
   useEffectWithDependencies(
-    "Mutation observer",
+    "PR files mutation observer",
     deps => ({ target: () => deps.all.mutationObserverTarget() }),
     deps => {
       observer.observe(deps.target.element, options);
@@ -200,4 +72,10 @@ function useMutationObserverEffect(observer: MutationObserver, options: Mutation
     },
     []
   );
+}
+
+function getFileExtension(container: ResolvedDomDependency) {
+  return getOriginalFilePath(container)
+    .split(".")
+    .pop()!;
 }
